// src/web/src/app/api/auth/models/verify-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MembershipResponse } from '../models/membership-response';
export interface VerifyResponse {
  address: string;
  expiresAt: string;
  holder: boolean;
  jwt: string;
  memberships: Array<MembershipResponse>;
  platformAdmin: boolean;
  userId: (string | null);
}